test: split unit and integration test lanes
