+++ src/main_test.go
@@ -1,3 +1,5 @@
+//go:build !integration
+
 package main
 
 import (
